uploading a new student profile picture will remove the existing in public

+++ app/Http/Controllers/StudentController.php
@@ -62,20 +62,29 @@ class StudentController extends Controller
      */
     public function update(Request $request, Student $student) 
     {
-        // custom image name
-        $new_profile_picture_name = time() . 
-                                    '-' . 
-                                    Auth::user()->student->last_name . 
-                                    '.' .
-                                    $request->profile_picture->extension();
+        if (isset($request->profile_picture))
+        {
+            // custom image name
+            $new_profile_picture_name = time() . 
+            '-' . 
+            Auth::user()->student->last_name . 
+            '.' .
+            $request->profile_picture->extension();
 
-        // if there is already a profile picture in public, delete it
-        if (File::exists(public_path('img/profile-pictures/' . $new_profile_picture_name))) {
+            // if there is already a profile picture in public, delete it
+            if (File::exists(public_path('img/profile-pictures/' . $new_profile_picture_name))) 
+            {
             File::delete(public_path('img/profile-pictures/' . $new_profile_picture_name));
+            }
+
+            // move the image file to public/img/profile-pictures
+            $request->profile_picture->move(public_path('img/profile-pictures'), $new_profile_picture_name);
+        } 
+        else 
+        {
+            $new_profile_picture_name = null;
         }
         
-        // move the image file to public/img/profile-pictures
-        $request->profile_picture->move(public_path('img/profile-pictures'), $new_profile_picture_name);
 
         $student->update([
             'first_name' => $request->first_name,
